Added the ability to reference custom effect classes from within the author's plugin or someone else's plugin.
1) CardEffect and CardTrait now have error checking for the name property. It first checks the configurations associated plugin assembly for the type.
If it is found it resolves it to the AssemblyQualifiedName so that the game is able to fetch and instantiate the provided effect class from another Assembly.
If it is not found then it tries searching the game's assembly file for the type. It will resolve to the name of the class if found. Otherwise its an error.
2) mod_reference allows one to search someone elses plugin for the type given a mod guid (use case a community provided plugin with many reusable custom effect classes for use).
This overrides the search mechanism to start looking at the provided mod for the type, then the base game.