For boolean properties the logic was defaulting them all to false. We don't need to AND with the value of checkOverride
If you are overriding then you'd want the value in the object (if it is not defined in the json then its a noop)
If you are not overriding then you get a clean object whose values are all set to their defaults (if it is not defined then you get the default value again a noop)

Add a note to endless stats. Its only for enemy units.